prototype import functionality for web ui
