+++ utils/__init__.py
@@ -1,7 +1,24 @@
 import datetime
 import math
+import unicodedata
+import re
 from collections.abc import Iterable
 from itertools import chain, zip_longest
+
+
+def slugify(value, allow_unicode=False):
+    """
+    Convert to ASCII if 'allow_unicode' is False. Convert spaces to hyphens.
+    Remove characters that aren't alphanumerics, underscores, or hyphens.
+    Convert to lowercase. Also strip leading and trailing whitespace.
+    """
+    value = str(value)
+    if allow_unicode:
+        value = unicodedata.normalize('NFKC', value)
+    else:
+        value = unicodedata.normalize('NFKD', value).encode('ascii', 'ignore').decode('ascii')
+    value = re.sub(r'[^\w\s-]', '', re.sub(r'[:=]', '-', value).strip().lower())
+    return re.sub(r'[\s]+', '-', value)
 
 
 def dtformat(dt):
